requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
av==13.1.0
bert-score==0.3.13
bitsandbytes==0.40.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.0
dill==0.3.8
docstring_parser==0.16
einops==0.8.0
et_xmlfile==2.0.0
evaluate==0.4.3
faiss-cpu==1.10.0
fastapi==0.115.4
ffmpy==0.4.0
filelock==3.16.0
fire==0.7.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gradio==4.44.1
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.24.7
idna==3.10
importlib_resources==6.4.5
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonargparse==4.32.1
kiwisolver==1.4.7
lightning @ git+https://github.com/Lightning-AI/lightning@7eafab98eb9877372ea406607794dd9a9b64e804
lightning-utilities==0.11.7
-e git+https://github.com/Lightning-AI/lit-llama@d513022842f7ee54c86595ce636d3133e35f8a8c#egg=lit_llama
-e git+https://github.com/hiyouga/LLaMA-Factory.git@707ff5a3703ead7b1d374d928f92ed8a128573f0#egg=llamafactory
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.5
orjson==3.10.11
packaging==24.1
pandas==2.2.2
peft==0.12.0
pillow==10.2.0
portalocker==3.1.1
protobuf==5.28.3
psutil==6.1.0
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.4
rouge-chinese==1.0.3
rouge_score==0.1.2
ruff==0.7.2
sacrebleu==2.5.1
sacremoses==0.1.1
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.14.1
semantic-version==2.10.0
sentence-transformers==3.4.1
sentencepiece==0.2.0
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
sniffio==1.3.1
sse-starlette==2.1.3
starlette==0.41.2
sympy==1.13.2
tabulate==0.9.0
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.20.3
tomlkit==0.12.0
torch==2.4.0+cu121
torchaudio==2.4.0+cu121
torchmetrics==1.4.2
torchvision==0.19.0+cu121
tqdm==4.66.5
transformers==4.46.1
triton==3.0.0
trl==0.9.6
typer==0.12.5
typeshed_client==2.7.0
typing_extensions==4.12.2
tyro==0.8.14
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.32.0
websockets==12.0
xxhash==3.5.0
yarl==1.11.1
zstandard==0.23.0
